+++ Sources/KeyboardKit/_Deprecated/AudioFeedback+Deprecated.swift
@@ -1,0 +1,8 @@
+public extension AudioFeedback {
+    
+    @available(*, deprecated, renamed: "player")
+    static var systemPlayer: SystemAudioPlayer {
+        get { player }
+        set { player = newValue }
+    }
+}